+++ examples/syntax_error.py
@@ -18,8 +18,8 @@ import sys
 # # Uncomment this as a block.
 # # This tests that errors before the first st call get caught.
 # def foo():
-     # EXPECTED: inline exception
-     # a = not_a_real_variable  # noqa: F821 pylint:disable=undefined-variable,unused-variable
+# EXPECTED: inline exception
+# a = not_a_real_variable  # noqa: F821 pylint:disable=undefined-variable,unused-variable
 
 # foo()