+++ frontend/src/App.tsx
@@ -41,6 +41,7 @@ import {
   ReportElement,
   SimpleElement,
 } from "lib/DeltaParser"
+import { setCookie } from "lib/utils"
 import {
   BackMsg,
   Delta,
@@ -142,7 +143,7 @@ export class App extends PureComponent<Props, State> {
       return this.connectionManager
         ? this.connectionManager.getBaseUriParts()
         : undefined
-    })
+    }, true)
     this.elementListBufferTimerIsSet = false
     this.elementListBuffer = null
 
@@ -225,6 +226,8 @@ export class App extends PureComponent<Props, State> {
       )
       this.widgetMgr.sendUpdateWidgetsMessage()
       this.setState({ dialog: null })
+    } else {
+      setCookie("_xsrf", "")
     }
   }
 
@@ -435,15 +438,10 @@ export class App extends PureComponent<Props, State> {
 
     document.title = `${reportName} · Streamlit`
 
+    MetricsManager.current.setReportHash(newReportHash)
     MetricsManager.current.clearDeltaCounter()
 
-    MetricsManager.current.enqueue("updateReport", {
-      // Create a hash that uniquely identifies this "project" so we can tell
-      // how many projects are being created with Streamlit while still keeping
-      // possibly-sensitive info like the scriptPath outside of our metrics
-      // services.
-      reportHash: newReportHash,
-    })
+    MetricsManager.current.enqueue("updateReport")
 
     if (reportHash === newReportHash) {
       this.setState({