+++ lib/tests/streamlit/Server_test.py
@@ -177,7 +177,6 @@ class ServerTest(ServerTestCase):
             # present in the response from the server.
             self.assertIsNone(ws_client.headers.get("Sec-Websocket-Extensions"))
 
-
     @tornado.testing.gen_test
     def test_forwardmsg_hashing(self):
         """Test that outgoing ForwardMsgs contain hashes."""
@@ -461,6 +460,20 @@ class HealthHandlerTest(tornado.testing.AsyncHTTPTestCase):
         self._is_healthy = False
         response = self.fetch("/healthz")
         self.assertEqual(503, response.code)
+
+    def test_healthz_without_csrf(self):
+        config._set_option("server.enableXsrfProtection", False, "test")
+        response = self.fetch("/healthz")
+        self.assertEqual(200, response.code)
+        self.assertEqual(b"ok", response.body)
+        self.assertNotIn("Set-Cookie", response.headers)
+
+    def test_healthz_with_csrf(self):
+        config._set_option("server.enableXsrfProtection", True, "test")
+        response = self.fetch("/healthz")
+        self.assertEqual(200, response.code)
+        self.assertEqual(b"ok", response.body)
+        self.assertIn("Set-Cookie", response.headers)
 
 
 class PortRotateAHundredTest(unittest.TestCase):