+++ lib/tests/streamlit/bootstrap_test.py
@@ -85,12 +85,14 @@ class BootstrapPrintTest(unittest.TestCase):
         )
 
         with patch.object(config, "get_option", new=mock_get_option), patch.object(
-                config, "is_manually_set", new=mock_is_manually_set
+            config, "is_manually_set", new=mock_is_manually_set
         ):
             bootstrap._print_url(True)
 
         out = sys.stdout.getvalue()
-        self.assertTrue("Welcome to Streamlit. Check out our demo in your browser." in out)
+        self.assertTrue(
+            "Welcome to Streamlit. Check out our demo in your browser." in out
+        )
         self.assertTrue("URL: http://the-address" in out)
 
     def test_print_urls_configured(self):
@@ -209,7 +211,7 @@ class BootstrapPrintTest(unittest.TestCase):
             {"browser.serverAddress": False}
         )
         mock_get_option = testutil.build_mock_config_get_option(
-            {"server.headless": False, "server.port": 9988, "global.useNode": False}
+            {"server.headless": False, "server.port": 9988, "global.developmentMode": False}
         )
 
         mock_get_internal_ip.return_value = "internal-ip"
@@ -233,7 +235,7 @@ class BootstrapPrintTest(unittest.TestCase):
                 "server.headless": False,
                 "server.baseUrlPath": "foo",
                 "server.port": 8501,
-                "global.useNode": False,
+                "global.developmentMode": False,
             }
         )
 
@@ -258,7 +260,7 @@ class BootstrapPrintTest(unittest.TestCase):
                 "server.headless": False,
                 "server.baseUrlPath": "foo",
                 "server.port": 8501,
-                "global.useNode": False,
+                "global.developmentMode": False,
             }
         )