+++ lib/tests/streamlit/config_test.py
@@ -285,7 +285,6 @@ class ConfigTest(unittest.TestCase):
                 "global.showWarningOnDirectExecution",
                 "global.suppressDeprecationWarnings",
                 "global.unitTest",
-                "global.useNode",
                 "runner.magicEnabled",
                 "runner.installTracer",
                 "runner.fixMatplotlib",
@@ -297,12 +296,13 @@ class ConfigTest(unittest.TestCase):
                 "s3.region",
                 "s3.secretAccessKey",
                 "s3.url",
-                "server.enableCORS",
-                "server.enableWebsocketCompression",
                 "server.baseUrlPath",
+                "server.enableCORS",
                 "server.cookieSecret",
-                "server.folderWatchBlacklist",
+                "server.enableWebsocketCompression",
+                "server.enableXsrfProtection",
                 "server.fileWatcherType",
+                "server.folderWatchBlacklist",
                 "server.headless",
                 "server.liveSave",
                 "server.address",
@@ -350,6 +350,13 @@ class ConfigTest(unittest.TestCase):
             str(e.value),
             "server.port does not work when global.developmentMode is true.",
         )
+
+    def test_check_conflicts_server_csrf(self):
+        config._set_option("server.enableXsrfProtection", True, "test")
+        config._set_option("server.enableCORS", True, "test")
+        with patch("streamlit.config.LOGGER") as patched_logger:
+            config._check_conflicts()
+            patched_logger.warning.assert_called_once()
 
     def test_check_conflicts_browser_serverport(self):
         config._set_option("global.developmentMode", True, "test")